refactored code and fixed listener
