+++ api/go.mod
@@ -7,7 +7,4 @@ require (
 	github.com/lib/pq v1.10.9 // direct
 )
 
-require (
-	github.com/gorilla/mux v1.8.1 // indirect
-	github.com/rs/cors v1.11.0 // indirect
-)
+require github.com/gorilla/mux v1.8.1
